refactor: split remaining code into utils and inserter files
